fix: 快捷键注册修复
